Some additional work on __sent_get_request
Still doesn't produce result. For some reason, no answer packets are being received. We get a packet under "1 remaining", but that's just the packet we sent, returned back to us because it isn't answered. That's why its payload is blank; it's our own packet, we didn't include a payload when we sent it.